lazy header and bai file initialization
There're no support for shared libraries yet, and bindings
are going to call CLI tools for every operation. Under such
circumstances it's crucial to not do unnecessary stuff in
BamFile constructor.
One of observations is that SAM header is unused for viewing reads,
and therefore it's parsing can be delayed until necessary.
As for BAI initialization, that wasn't absolutely necessary as
there're two common cases. Either user iterates ALL alignments in
a file, and time spent on reading BAI is just negligible, or he/she
makes a region query, and we must read index file anyway.