Finally, GDC goes to hell!
LDMD2 takes its place for building releases.
The reason is that I can't release totally unreliable software!

I built sambamba with all three compilers today and found out that
sambamba-merge crashed when built with GDC, but it works with the
other two compilers.
Then I remembered that in the summer, I had similar issues with
sambamba-index but kind of solved them by reordering list of files
passed to GDC. I thought then that maybe it's a good idea to check
how merge tool behaved over time, and used git bisect for that.
Well, in the middle of bisecting process I realized that output of
GDC was always fucked up! In one of runs the tool exited before it
finished merging, in the other it finished successfully.

Yet another reason to stop using GDC for this project is that
multithreading no longer works with 4.8 branch. What disturbs me
even more is the fact that I filed this bug on GDC bugzilla a long
ago and still don't have any response. So I use old 4.7 branch
which is based on 2.059 version of D, and even this more stable
branch miscompiles my code.
Even though it will cost me some performance, I will use LDC2
from now on for building releases. At least its developers are
much more active and friendly (and there are more than 1 of them).