Prevent output queue from growing indefinitely when output is slow
Allowing worker threads to offload completed chunks to an output thread
raises the possibility that a slow output (slow device, piped into slow
bgzip etc.) can grow the queue indefinitely. If this continues for large
input then memory usage will grow extremely large.

It is possible to eliminate this effect by reducing the number of worker
threads or improving the output speed (e.g. pbgzip) but it's better to
log and prevent excessive queueing than to allow it to continue
happening indefinitely.

This change caps the number of queued chunks to 2 * number of threads.