make error message more informative
When one passes a float, let's say -m=2.5G the only message that is returned is: "sambamba-sort: couldn't parse ", which is not very informative, as this someone may blame the input BAM file and spend some time trying to figure out what is wrong with it. The updated  error message points directly into the point saving users time.